Refactored to package-style
