+++ CV/closedcv/setup.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+
+from distutils.core import setup
+
+setup(name='closedcv',
+      version='1.0',
+      description='Python Distribution Utilities',
+      author='Alex Karavaev',
+      author_email='alexkaravev@gmail.com',
+      packages=['closedcv'],
+     )
